+++ src/app.tsx
@@ -1,26 +1,34 @@
 import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from './service/queryClient';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-//Components
-import Home from './pages';
-import Header from './layouts/header';
-import Footer from './layouts/footer';
+import { queryClient } from './services/query-client';
+import {
+  createBrowserRouter,
+  createRoutesFromElements,
+  Route,
+  RouterProvider,
+} from 'react-router-dom';
 //Pages
-import TvShows from './pages/tvShows';
+import Home from './pages';
+import TvShows from './pages/tv-shows';
 import Movies from './pages/movies';
+import Error404 from './errors/error-404';
+//Layout
+import Root from './layouts/root';
+
+const router = createBrowserRouter(
+  createRoutesFromElements(
+    <Route path='/' element={<Root />}>
+      <Route index element={<Home />} />
+      <Route path='tv-shows' element={<TvShows />} />
+      <Route path='movies' element={<Movies />} />
+      <Route path='*' element={<Error404 />} />
+    </Route>
+  )
+);
 
 const App = () => {
   return (
     <QueryClientProvider client={queryClient}>
-      <BrowserRouter>
-        <Header />
-        <Routes>
-          <Route path='/' element={<Home />} />
-          <Route path='/tv-shows' element={<TvShows />} />
-          <Route path='/movies' element={<Movies />} />
-        </Routes>
-        <Footer />
-      </BrowserRouter>
+      <RouterProvider router={router} />
     </QueryClientProvider>
   );
 };
