+++ src/errors/error-404.tsx
@@ -1,0 +1,5 @@
+const Error404 = () => {
+  return <div>Error404 NOT FOUND</div>;
+};
+
+export default Error404;
