+++ src/errors/error-page.tsx
@@ -1,0 +1,9 @@
+const ErrorPage = () => {
+  return (
+    <>
+      <p>Ups... Something happened</p>
+    </>
+  );
+};
+
+export default ErrorPage;
